requirements.txt
streamlit==1.36.0
opencv-python-headless==4.6.0.66
numpy==1.26.4
Pillow==10.2.0
matplotlib==3.7.3

paddleocr==2.7.0.0
paddlepaddle==2.5.2

torch==2.2.2
torchvision==0.17.2
vietocr==0.3.13
